src/D_coref/img_llm_extract_coref_files/requirements.txt
aiohappyeyeballs==2.4.0; python_version >= '3.8'
aiohttp==3.10.5; python_version >= '3.8'
aiosignal==1.3.1; python_version >= '3.7'
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.4.0; python_version >= '3.8'
argon2-cffi==23.1.0; python_version >= '3.7'
argon2-cffi-bindings==21.2.0; python_version >= '3.6'
arrow==1.3.0; python_version >= '3.8'
asttokens==2.4.1
async-lru==2.0.4; python_version >= '3.8'
attrs==24.2.0; python_version >= '3.7'
babel==2.16.0; python_version >= '3.8'
beautifulsoup4==4.12.3; python_full_version >= '3.6.0'
bleach==6.1.0; python_version >= '3.8'
certifi==2024.8.30; python_version >= '3.6'
cffi==1.17.1; python_version >= '3.8'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
comm==0.2.2; python_version >= '3.8'
dataclasses-json==0.6.7; python_version >= '3.7' and python_version < '4.0'
debugpy==1.8.5; python_version >= '3.8'
decorator==5.1.1; python_version >= '3.5'
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
distro==1.9.0; python_version >= '3.6'
executing==2.1.0; python_version >= '3.8'
fastapi==0.115.0; python_version >= '3.8'
fastjsonschema==2.20.0
fqdn==1.5.1
frozenlist==1.4.1; python_version >= '3.8'
greenlet==3.1.0; python_version < '3.13' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.5; python_version >= '3.8'
httpx==0.27.2; python_version >= '3.8'
idna==3.10; python_version >= '3.6'
ipykernel==6.29.5; python_version >= '3.8'
ipython==8.27.0; python_version >= '3.10'
ipywidgets==8.1.5; python_version >= '3.7'
isoduration==20.11.0
jedi==0.19.1; python_version >= '3.6'
jinja2==3.1.4; python_version >= '3.7'
jiter==0.5.0; python_version >= '3.8'
json5==0.9.25; python_version >= '3.8'
jsonpatch==1.33; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
jsonpointer==3.0.0; python_version >= '3.7'
jsonschema[format-nongpl]==4.23.0; python_version >= '3.8'
jsonschema-specifications==2023.12.1; python_version >= '3.8'
langchain==0.3.0; python_version >= '3.9' and python_version < '4.0'
langchain-community==0.3.0; python_version >= '3.9' and python_version < '4.0'
langchain-core==0.3.1; python_version >= '3.9' and python_version < '4.0'
langchain-openai==0.2.0; python_version >= '3.9' and python_version < '4.0'
langchain-text-splitters==0.3.0; python_version >= '3.9' and python_version < '4.0'
langsmith==0.1.121; python_version < '4.0' and python_full_version >= '3.8.1'
markupsafe==2.1.5; python_version >= '3.7'
marshmallow==3.22.0; python_version >= '3.8'
matplotlib-inline==0.1.7; python_version >= '3.8'
mistune==3.0.2; python_version >= '3.7'
multidict==6.1.0; python_version >= '3.8'
mypy-extensions==1.0.0; python_version >= '3.5'
nbclient==0.10.0; python_full_version >= '3.8.0'
nbconvert==7.16.4; python_version >= '3.8'
nbformat==5.10.4; python_version >= '3.8'
nest-asyncio==1.6.0; python_version >= '3.5'
notebook==7.2.2; python_version >= '3.8'
notebook-shim==0.2.4; python_version >= '3.7'
numpy==1.26.4; python_version >= '3.12'
openai==1.46.0; python_full_version >= '3.7.1'
orjson==3.10.7; python_version >= '3.8'
overrides==7.7.0; python_version >= '3.6'
packaging==24.1; python_version >= '3.8'
pandas==2.2.2; python_version >= '3.9'
pandocfilters==1.5.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
parso==0.8.4; python_version >= '3.6'
pexpect==4.9.0; sys_platform != 'win32' and sys_platform != 'emscripten'
platformdirs==4.3.6; python_version >= '3.8'
prometheus-client==0.20.0; python_version >= '3.8'
prompt-toolkit==3.0.47; python_full_version >= '3.7.0'
psutil==6.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==17.0.0; python_version >= '3.8'
pycparser==2.22; python_version >= '3.8'
pydantic==2.9.2; python_version >= '3.8'
pydantic-core==2.23.4; python_version >= '3.8'
pydantic-settings==2.5.2; python_version >= '3.8'
pygments==2.18.0; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.1; python_version >= '3.8'
python-json-logger==2.0.7; python_version >= '3.6'
pytz==2024.2
pyyaml==6.0.2; python_version >= '3.8'
pyzmq==26.2.0; python_version >= '3.7'
referencing==0.35.1; python_version >= '3.8'
regex==2024.9.11; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
rfc3339-validator==0.1.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
rfc3986-validator==0.1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
rpds-py==0.20.0; python_version >= '3.8'
send2trash==1.8.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
setuptools==75.1.0; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.1; python_version >= '3.7'
soupsieve==2.6; python_version >= '3.8'
sqlalchemy==2.0.35; python_version >= '3.7'
stack-data==0.6.3
starlette==0.38.5; python_version >= '3.8'
tenacity==8.5.0; python_version >= '3.8'
terminado==0.18.1; python_version >= '3.8'
tiktoken==0.7.0; python_version >= '3.8'
tinycss2==1.3.0; python_version >= '3.8'
tornado==6.4.1; python_version >= '3.8'
tqdm==4.66.5; python_version >= '3.7'
traitlets==5.14.3; python_version >= '3.8'
types-python-dateutil==2.9.0.20240906; python_version >= '3.8'
typing-extensions==4.12.2; python_version >= '3.8'
typing-inspect==0.9.0
tzdata==2024.1; python_version >= '2'
uri-template==1.3.0
urllib3==2.2.3; python_version >= '3.8'
uvicorn==0.30.6; python_version >= '3.8'
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0; python_version >= '3.8'
widgetsnbextension==4.0.13; python_version >= '3.7'
yarl==1.11.1; python_version >= '3.8'
